requirements.txt
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
mechanicalsoup>=1.3.0
typer[all]>=0.12.0
rich>=13.0.0
pyyaml>=6.0
requests>=2.31.0
playwright>=1.44.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.9
websockets>=12.0
